+++ src/engine/timer.rs
@@ -2,7 +2,7 @@ use std::time::Instant;
 
 pub struct Timer {
     start_time: Instant,
-    adjust: f32,
+    adjust: f64,
 }
 
 #[allow(dead_code)]
@@ -18,15 +18,15 @@ impl Timer {
         self.start_time = Instant::now();
     }
 
-    pub fn elapsed(&self) -> f32 {
+    pub fn elapsed(&self) -> f64 {
         self.true_elapsed() + self.adjust
     }
 
-    pub fn true_elapsed(&self) -> f32 {
-        self.start_time.elapsed().as_millis() as f32 * 0.001
+    pub fn true_elapsed(&self) -> f64 {
+        self.start_time.elapsed().as_millis() as f64 * 0.001
     }
 
-    pub fn forward(&mut self, seconds: f32) {
+    pub fn forward(&mut self, seconds: f64) {
         self.adjust = (self.adjust + seconds).max(-self.true_elapsed())
     }
 }
